Memory Management Utilities for Monte Carlo Simulations

Functions to help manage memory usage during large Monte Carlo simulations, including data summarization, streaming results, and memory monitoring. Create a summary function for data reduction

Usage

create_data_summarizer(stats = c("mean", "sd", "quantiles"), vars = NULL)

Arguments

stats

Character vector of statistics to compute

vars

Character vector of variables to summarize (NULL for all)

Value

Function that summarizes data

Examples

# Create a summarizer that computes means and SDs
summarizer <- create_data_summarizer(
  stats = c("mean", "sd", "quantiles"),
  vars = c("y", "a")
)

# Use in Monte Carlo
if (FALSE) { # \dontrun{
results <- margot_monte_carlo(
  n_reps = 1000,
  n_per_rep = 100,
  dgp_params = list(
    waves = 2,
    treatments = "a",
    interventions = list(
      natural = function(data, time, trt) data[[trt]],
      shifted = function(data, time, trt) pmin(data[[trt]] + 1, 2)
    )
  ),
  estimator_fn = function(data) list(estimate = mean(data$t2_y)),
  summarize_fn = summarizer
)
} # }