Create multi-treatment shift function (LMTP style)

Creates shift functions that can modify multiple treatments simultaneously, compatible with LMTP's multi-treatment framework.

Usage

create_multi_treatment_shift(shift_logic, start_wave = 1)

Arguments

shift_logic

Function that takes (data, trt_list) and returns list of shifted values for each treatment

start_wave

Wave at which to start shifting

Value

Function suitable for LMTP multi-treatment interventions

Examples

# Shift both treatments to 1
both_high <- create_multi_treatment_shift(
  shift_logic = function(data, trt_list) {
    lapply(unlist(trt_list), function(trt) {
      ifelse(!is.na(data[[trt]]), 1, NA)
    })
  }
)

# Complex interaction: shift based on variable names
interaction <- create_multi_treatment_shift(
  shift_logic = function(data, trt_list) {
    all_trts <- unlist(trt_list)
    shifted <- list()
    
    for (trt in all_trts) {
      if (grepl("treatment_a", trt)) {
        shifted[[trt]] <- ifelse(data[[trt]] != 1, 1, data[[trt]])
      } else if (grepl("treatment_b", trt)) {
        shifted[[trt]] <- ifelse(data[[trt]] != 1, 1, data[[trt]])
      }
    }
    shifted
  }
)